Fix Ctrl-Command-Power reboot key combo
After the reboot key combo is held down the Mac OS does a talk reg 2
continuously until the modifier keys are lifted. The talk reg 2 wasn't
being updated during that loop so system would start.

Now reg 2 is checked right before it is sent over the ADB bus so
the modifier keys are always current.

Another fix was done with the power button. The power key up state
was malformed so the power button was stuck down after pressing it.
This has been fixed.

Removed code related to double queuing the power button on the USB side.
The double queuing is done now on the ADB side. The power button
takes up both key press packets so it need to be queued twice and
sent on the same talk register 0.