Add Command and Option key swapping
The Command and Option key are now swappable with the `K` special
command key. The swappable option can be saved to flash so it
doesn't need to changed every time the QuokkADB if powered.

Added `Ctrl + Command + Option + [key]` as alternate way to send special
commands. This was added because some keyboards don't allow
`Caps lock + Command + Shift + [key]` as a valid key press.