Advanced keyboard and mouse features working
< CTRL + ALT + COMMAND + P > now prints all the adjustable settings
for the keyboard and mouse along with their shortcuts.

Mouse wheel introduced, though it only works on mice that support
the mouse wheel in HID boot protocol mode. Added the ability to switch
axis direction with the mouse wheel and the ability to set the
sensitivity of the scrolling. The mouse wheel sends up and down arrow
key presses.