Power button working
Currently INSERT, F15, and HELP buttons activate the power button line
on the NeXT keyboard interface. If the QuokKM5 is powered via the
USB-C port, they keyboard can power on the NeXT. If it isn't, the power
button on the QuokKM5 must be pressed. Once the NeXT is on the keyboard
can turn "press" the power button no matter how the QuokKM5 is powered.