Update to latest TinyUSB and the RP platform
Using the latest as of this commit, PlatformIO Raspberry Pi platform
and custom spin of the arduino-pico package with the pico-sdk 2.0.0
and the 0.17.0 TinyUSB library.

This seems to fix a corner case where phantom click were occurring
when two keyboard were hooked up along with a mouse on one of the
keyboard's hub ports.